rmdirRecursiveSync DRY
