+++ test/acceptance/ported/server.js
@@ -1,25 +1,11 @@
-require('../../support/test_helper');
+var testHelper = require('../../support/test_helper');
 
 var assert = require('../../support/assert');
-var fs = require('fs');
 var cartodbServer = require('../../../lib/cartodb/server');
 var ServerOptions = require('./support/ported_server_options');
 var testClient = require('./support/test_client');
 
 var BaseController = require('../../../lib/cartodb/controllers/base');
-
-function rmdir_recursive_sync(dirname) {
-  var files = fs.readdirSync(dirname);
-  for (var i=0; i<files.length; ++i) {
-    var f = dirname + "/" + files[i];
-    var s = fs.lstatSync(f);
-    if ( s.isFile() ) {
-      fs.unlinkSync(f);
-    } else {
-        rmdir_recursive_sync(f);
-    }
-  }
-}
 
 describe('server', function() {
 
@@ -35,7 +21,7 @@ describe('server', function() {
     after(function() {
         BaseController.prototype.req2params = req2paramsFn;
 
-        rmdir_recursive_sync(global.environment.millstone.cache_basedir);
+        testHelper.rmdirRecursiveSync(global.environment.millstone.cache_basedir);
     });
 
     ////////////////////////////////////////////////////////////////////
