+++ test/acceptance/ported/server_gettile.js
@@ -1,7 +1,6 @@
-require('../../support/test_helper');
+var testHelper = require('../../support/test_helper');
 
 var assert = require('../../support/assert');
-var fs = require('fs');
 var mapnik = require('windshaft').mapnik;
 var semver = require('semver');
 var cartodbServer = require('../../../lib/cartodb/server');
@@ -9,20 +8,6 @@ var ServerOptions = require('./support/ported_server_options');
 var testClient = require('./support/test_client');
 
 var BaseController = require('../../../lib/cartodb/controllers/base');
-
-function rmdir_recursive_sync(dirname) {
-  var files = fs.readdirSync(dirname);
-  for (var i=0; i<files.length; ++i) {
-    var f = dirname + "/" + files[i];
-    var s = fs.lstatSync(f);
-    if ( s.isFile() ) {
-      fs.unlinkSync(f);
-    }
-    else {
-        rmdir_recursive_sync(f);
-    }
-  }
-}
 
 describe('server_gettile', function() {
 
@@ -41,7 +26,7 @@ describe('server_gettile', function() {
     after(function() {
         BaseController.prototype.req2params = req2paramsFn;
 
-        rmdir_recursive_sync(global.environment.millstone.cache_basedir);
+        testHelper.rmdirRecursiveSync(global.environment.millstone.cache_basedir);
     });
 
     function imageCompareFn(fixture, done) {
